refactor: extract XML parsers into speaker-xml.js
Split api.js along its natural seam: REST transport stays in api.js,
all parseXxxXml/parseXxxEl pairs move to a new speaker-xml.js module.
Both forms remain fixture-testable and are consumed by both the REST
fetchers (in api.js) and the WS dispatch path (in speaker-state.js).

Renames admin/test/test_api_parsers.js → test_speaker_xml.js and
re-points its imports; updates speaker-state.js to import parsers from
speaker-xml.js while keeping fetcher imports from api.js. XML builders
(xmlEscape, buildZoneXml, inline string-template XML in postX
functions) intentionally stay in api.js — this module is read-only.
